moved URL status check to config validation fixed heartbeat status issues
